implement error logging for New Relic
       Episode IV
      A  NEW  RELIC

There is unrest in the application.
Or at least, there might be, we
don't know, which is the problem,
which is why we started writing
logger modules with event handlers
that we can call from our global
error boundaries.

That way, whenever there's an
unhandled exception in an app, we
can log it to a location accessible
to the hosting provider, instead of
just the JS console, and if we've
been clever enough to provide
sourcemaps with our app, or to
the logger, then we can (hopefully)
solve the problem.

This one logs to New Relic. Con-
figure it according to the README.